camera/capture_image-manual.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset='utf-8'>
    <title>HTML Media Capture Test: capture image to produce one image file</title>
    <link rel='author' title='Intel' href='http://www.intel.com'>
    <link rel='help' href='http://www.w3.org/TR/html-media-capture/#security'>
    <link rel='help' href='http://www.w3.org/TR/html5/forms.html#file-upload-state-(type=file)'>
    <meta name='flags' content='interact'>
    <script src='../support/testharness.js'></script>
    <script src='../support/testharnessreport.js'></script>
  </head>
  <body>
    <p>This test validates that the user agent need seek user consent before initiating capture of content by camera, and correctly produces image file when the user goes ahead with a capture.</p>

    <p>Test steps:</p>
    <ol>
      <li>Click the button. If prompted for permission, please allow.</li>
      <li>Capture a image.</li>
    </ol>

    <form>
      <input type="file" accept="image/*" capture>
    </form>

    <div id='log'></div>

    <script>
      setup({explicit_done: true, explicit_timeout: true});

      var input = document.querySelector('input[type=file]');
      input.onchange = function() {
        test(function() {
          var files = input.files;
          assert_equals(files.length, 1, 'Capture one image file');
          assert_regexp_match(files[0].type, /^image\//, 'The capture image type is image/*');
        }, document.title);

        done();
      }
    </script>
  </body>
</html>
